backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
requests==2.32.3
environs==14.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-slugify==8.0.4